[*] linux

+++ core/src/source/model/Poetry.cpp
@@ -59,7 +59,7 @@ bool lifuren::Poetry::matchLabel() {
         LabelText& labelRef = iterator->second;
         // TODO: 验证词能否正确匹配或者添加词牌
         if(
-            labelRef.segmentSize == this->simpleParagraphs.size() &&
+            labelRef.segmentSize == static_cast<int>(this->simpleParagraphs.size()) &&
             labelRef.segmentRule == segmentRule
         ) {
             this->label = &labelRef;
